TLDR-531 test start fixed
* path_detect forword fix
* some debug_mode and path_debug bugs fixed

+++ dedoc/readers/pdf_reader/pdf_image_reader/ocr/ocr_cell_extractor.py
@@ -82,8 +82,9 @@ class OCRCellExtractor:
         for tree_node in tree_table_nodes:
             x_coord = space
             cell_image = BBox.crop_image_by_box(src_image, tree_node.crop_text_box)
-            image_path = os.path.join(self.config.get("path_debug"), "debug_tables", "batches", "cell_croped.png")
-            cv2.imwrite(image_path, cell_image)
+            if self.config.get("debug_mode", False) and self.config.get("path_debug", False) and os.path.exists(self.config.get("path_debug")):
+                image_path = os.path.join(self.config.get("path_debug"), "debug_tables", "batches", "cell_croped.png")
+                cv2.imwrite(image_path, cell_image)
             cell_height, cell_width = cell_image.shape[0], cell_image.shape[1]
 
             stacked_image[y_prev:y_prev + cell_height, x_coord:x_coord + cell_width] = cell_image
